+++ google/CHANGELOG.md
@@ -4,7 +4,7 @@ This file documents changes to the `workbench-google` library, including notes o
 
 ## 0.18
 
-SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-util" % "0.18-38342fa"`
+SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-util" % "0.18-TRAVIS-REPLACE-ME"`
 
 ### Changed
 - Moved `SamModel`
@@ -14,6 +14,7 @@ SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-util" % "0.18
 - Add `enableService` function to `GoogleProjectDAO`
 - removing member from google group handles 400 error when user does not exist
 - Add `getAncestry` to HttpGoogleProjectDAO
+- Add `createProject(projectName: String, parentId: String, parentType: GoogleParentResourceType)`
 
 ## 0.17
 - upgrade cats to 1.4.0 and scala to 2.12.7